Upgrade Clojure+Java dependencies
I personally tend to advocate for keeping deps current as a general
practice, but in this case I am specifically motivated to upgrade
stylefruits/gniazdo because it has an upgraded version of
org.eclipse.jetty.websocket/websocket-client which includes a fix for
[issue #1262][1] which was causing my app to print out WARNING log
messages when I use *this* library.

All tests passed when I ran `lein test`.

[1]: https://github.com/eclipse/jetty.project/issues/1262