Log when Websocket connection closes or errors
I maintain [a tool][1] that uses this library and I recently had to
debug an issue wherein the tool would hang in certain circumstances. It
turned out that I was executing a JS expression that was returning too
much data — I didn’t know about [the default 1MB limit][2] and it took
me a good while to figure out what was happening. If this library had
logged something when the connection errored and/or closed then it would
have made debugging this problem easier and faster. So I thought I’d
contribute that enhancement.

[1]: https://fundingcircle.github.io/fc4-framework/tool/
[2]: https://git.io/fjDeX