Adopt lein-tools-deps
Before this the dependencies were declared redundantly in two different
places (`project.clj` and `deps.edn`) and these two declarations had to
be kept in sync manually — which was bound to fail (via human error)
sooner or later.

lein-tools-deps allows us to specify the dependencies in one place while
ensuring that both tools.deps and Leiningen will continue to work with
this project.

https://github.com/RickMoynihan/lein-tools-deps